API restructuration
