+++ pom.xml
@@ -65,6 +65,12 @@
                 </exclusion>
             </exclusions>
         </dependency>
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter-api</artifactId>
+            <version>5.10.1</version>
+            <scope>test</scope>
+        </dependency>
 
         <!-- Autres dépendances nécessaires pour votre projet -->
     </dependencies>
