+++ src/main/java/com/jarspeed/api/config/package-info.java
@@ -1,7 +1,8 @@
 /**
  * Configuration classes for the JarSpeed API application.
  * <p>
- * This package contains configuration-related classes, such as security configuration
+ * This package contains configuration-related classes, such as security
+ * configuration
  * and any other Java-based configuration for the Spring Boot application.
  * </p>
  */
