--- src/main/java/com/jarspeed/api/controller/package-info.java
@@ -1,9 +1,0 @@
-/**
- * Controller classes for handling API requests in the JarSpeed API application.
- * <p>
- * This package contains controllers that handle HTTP requests and responses.
- * They act as the intermediaries between the HTTP requests from the client and
- * the business logic handled in the service layer.
- * </p>
- */
-package com.jarspeed.api.controller;
