+++ src/main/java/com/jarspeed/api/course/Course.java
@@ -1,0 +1,19 @@
+package com.jarspeed.api.course;
+
+/**
+ * Entity representing a sports course in the system.
+ * <p>
+ * This class is a JPA entity that defines the structure of the
+ * table in the database. It includes information about the
+ * course such as its distance, duration, and the path traveled.
+ * </p>
+ */
+//@Entity
+public class Course {
+
+    /**
+     * id of the course.
+     */
+    private Long id;
+    // other attributes, getters and setters
+    }
