+++ src/main/java/com/jarspeed/api/course/CourseController.java
@@ -1,0 +1,42 @@
+package com.jarspeed.api.course;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+/**
+ * Controller class for managing courses.
+ * <p>
+ * This class handles HTTP requests and responses for operations related to
+ * courses.
+ * </p>
+ */
+@RestController
+@RequestMapping("/parcours")
+public class CourseController {
+    // Dependency injection for the Parcours service
+    /**
+     * Service to manage route operations.
+     * The Spring framework automatically injects an instance of
+     * ParcoursService.
+     */
+    @Autowired
+    private CourseService courseService;
+
+    /**
+     * Retrieves and returns a list of all routes.
+     * This method handles GET requests to obtain a list of all
+     * courses.
+     * It uses the ParcoursService to query the database and
+     * Returns the list of courses.
+     *
+     * @return The list of routes available in the database.
+     */
+    @GetMapping
+    public List<Course> getAllParcours() {
+        return courseService.findAll();
+    }
+}
