+++ src/main/java/com/jarspeed/api/course/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * The Course package contains all classes related to the management of courses.
+ * <p>
+ * This package includes classes for the controller, service, repository,
+ * and model (entity) of courses. It is responsible for the business logic,
+ * data management, and presentation of courses in the application.
+ * </p>
+ */
+package com.jarspeed.api.course;
