--- src/main/java/com/jarspeed/api/model/Parcours.java
@@ -1,19 +1,0 @@
-package com.jarspeed.api.model;
-
-/**
- * Entité représentant un parcours sportif dans le système.
- * <p>
- * Cette classe est une entité JPA qui définit la structure de la table des
- * parcours dans la base de données. Elle inclut des informations sur le
- * parcours comme sa distance, sa durée, et le chemin parcouru.
- * </p>
- */
-//@Entity
-public class Parcours {
-
-    /**
-     * id du parcours.
-     */
-    private Long id;
-    // autres attributs, getters et setters
-}
