--- src/main/java/com/jarspeed/api/model/User.java
@@ -1,14 +1,0 @@
-package com.jarspeed.api.model;
-
-/**
- * Entité représentant un utilisateur dans le système de suivi de
- * parcours sportifs.
- * <p>
- * Cette classe est une entité JPA qui définit la structure de la
- * table des utilisateurs dans la base de données.
- * </p>
- */
-//@Entity
-public class User {
-
-}
