--- src/main/java/com/jarspeed/api/model/package-info.java
@@ -1,8 +1,0 @@
-/**
- * Model classes representing entities in the JarSpeed API application.
- * <p>
- * This package contains the entity classes that model the application's
- * data structure and map to the database tables.
- * </p>
- */
-package com.jarspeed.api.model;
