--- src/main/java/com/jarspeed/api/repository/package-info.java
@@ -1,8 +1,0 @@
-/**
- * Repository classes for data access in the JarSpeed API application.
- * <p>
- * Repositories are interfaces that provide data access operations, such as CRUD
- * operations, to interact with the database using Spring Data JPA.
- * </p>
- */
-package com.jarspeed.api.repository;
