--- src/main/java/com/jarspeed/api/service/package-info.java
@@ -1,8 +1,0 @@
-/**
- * Service classes that contain business logic for the JarSpeed API application.
- * <p>
- * Services in this package process the business logic of the application and
- * call on the data access objects to interact with the database.
- * </p>
- */
-package com.jarspeed.api.service;
