+++ src/main/java/com/jarspeed/api/user/User.java
@@ -1,0 +1,15 @@
+package com.jarspeed.api.user;
+
+/**
+ * Entity representing a user in the system for tracking
+ * tracking system.
+ * <p>
+ * This class is a JPA entity that defines the structure of the * user table
+ * in the database.
+ * user table in the database.
+ * </p>
+ */
+//@Entity
+public class User {
+
+}
