+++ src/main/java/com/jarspeed/api/user/UserController.java
@@ -1,0 +1,27 @@
+package com.jarspeed.api.user;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * REST controller for user management.
+ * <p>
+ * This controller handles HTTP requests for operations on.
+ * users, such as retrieving, creating, updating and deleting
+ * users.
+ * </p>
+ */
+@RestController
+@RequestMapping("/users")
+public class UserController {
+
+    // Dependency injection for the User service
+    /**
+     * Service to manage user-related operations.
+     * The Spring framework automatically injects a UserService instance.
+     */
+    @Autowired
+    private UserService userService;
+    // Other termination points
+}
