+++ src/main/java/com/jarspeed/api/user/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * The User package groups classes related to user management.
+ * <p>
+ * This package comprises controllers, services, repositories,
+ * and models (entities) for users. It handles operations related to users,
+ * such as registration, authentication, and updating profile information.
+ * </p>
+ */
+package com.jarspeed.api.user;
