+++ src/test/java/com/jarspeed/api/JarSpeedApplicationTest.java
@@ -6,8 +6,5 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class JarSpeedApplicationTest {
 
-    @Test
-    void contextLoads() {
-    }
 
 }
