+++ src/test/java/com/jarspeed/api/repository/CourseRepositoryTest.java
@@ -1,10 +1,10 @@
-package com.jarspeed.api.controller;
+package com.jarspeed.api.repository;
 
 import org.junit.jupiter.api.BeforeEach;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class ParcoursControllerTest {
+class CourseRepositoryTest {
 
     @BeforeEach
     void setUp() {
