+++ src/test/java/com/jarspeed/api/service/CourseServiceTest.java
@@ -1,10 +1,10 @@
-package com.jarspeed.api.repository;
+package com.jarspeed.api.service;
 
 import org.junit.jupiter.api.BeforeEach;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class ParcoursRepositoryTest {
+class CourseServiceTest {
 
     @BeforeEach
     void setUp() {
